requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
pip==23.0.1
setuptools==65.5.0
